Extracted assistant and Parameters
