+++ ekspedientki/include/shop.h
@@ -1,6 +1,15 @@
 #ifndef SHOP_H
 #define SHOP_H
 
+
+
+#include <assert.h>
+#include <stdio.h>
+#include <stdlib.h>
+#include <pthread.h>
+#include <unistd.h>
+
+
 int zso();
 
 #endif
